pyproject.toml
[project]
name = "sfmpe"
description = "Structured Flow Matching for Posterior Estimation"
authors = [{name = "Giovanni Charles", email = "giovanni.charles@gmail.com"}]
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
]
requires-python = ">=3.10"
dependencies = [
    "arviz>=0.17.1",
    "matplotlib>=3.6.2",
    "optax>=0.1.3",
    "seaborn>=0.12.2",
    "tensorflow==2.15.0",
    "tensorflow-probability>=0.20.0",
    "tqdm>=4.64.1",
    "flax>=0.10.2",
    "sbijax==0.3.3.post1"
]
dynamic = ["version"]

[project.urls]
Homepage = "https://github.com/giovannic/sfmpe"

[project.optional-dependencies]
dev = [
    "pyright",
    "jaxtyping"
]
